Separate simple camera and camera_transformations example
